Use a sqlite DB to improve dashboard performance.
This PR combines fetch and data into a single file that essentially
performs ETL into a sqlite DB. The dashboard is then configured to use
the sqlite DB instead of the original CSV files.
