+++ requirements.txt
@@ -1,6 +1,5 @@
 black
 humanize
-matplotlib
 pandas
 requests
 streamlit
